Enable ability to multi-stage build ARM
This commit enables the underlying CMake infrastructure to handle
multi-stage ARM builds.

We take advantage of this in order to circumvent the seL4 build chain where
it requires a built root-task image in order to build/link the elfloader-tool
with the kernel and root-task image.

On the first build, the elfloader-tool is not aware of the root-task image because it
doesn't exist yet. On further invocations, you can set the environment variable
`FEL4_ROOT_TASK_IMAGE` to have it be picked up by the elfloader-tool and become
linked into the final bootable image.

This is not a production friendly feature, and will become solidified as
we determine how to best handle our build staging.