Enable verification builds, remove cmake restriction
Now that we have all options declared in toml, we do not need to rely too much on
any of the helper functions in the Cmake files.

This change removes the call to Cmake helper function `ApplyCommonReleaseVerificationSettings()`.

The Cmake crate will handle the `CMAKE_BUILD_TYPE` depending on the build profile.

Verification options can be set in the fel4.toml file.

Closes #25